apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
ultralytics==8.2.87
opencv-python-headless==4.10.0.84
supervision==0.22.0
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
reportlab==4.2.2
Pillow==10.4.0
python-dotenv==1.0.1
pydantic==2.8.2
aiofiles==23.2.1
python-magic==0.4.27
pydantic-settings==2.4.0
